+++ fresh.gen.ts
@@ -21,11 +21,11 @@ import * as $$7 from "./islands/SearchControls.tsx";
 import * as $$8 from "./islands/Slider.tsx";
 import * as $$$0 from "./sections/Button.story.tsx";
 import * as $$$1 from "./sections/Carousel.tsx";
-import * as $$$2 from "./sections/CollectionHighlights.tsx";
-import * as $$$3 from "./sections/Footer.tsx";
-import * as $$$4 from "./sections/Head.tsx";
-import * as $$$5 from "./sections/Header.tsx";
-import * as $$$6 from "./sections/HighlightedFeatures.tsx";
+import * as $$$2 from "./sections/Footer.tsx";
+import * as $$$3 from "./sections/Head.tsx";
+import * as $$$4 from "./sections/Header.tsx";
+import * as $$$5 from "./sections/HighlightedFeatures.tsx";
+import * as $$$6 from "./sections/Highlights.tsx";
 import * as $$$7 from "./sections/MarcasConvidadas.tsx";
 import * as $$$8 from "./sections/Newsletter.tsx";
 import * as $$$9 from "./sections/ProductDetails.tsx";
@@ -69,11 +69,11 @@ const manifest: DecoManifest = {
   sections: {
     "./sections/Button.story.tsx": $$$0,
     "./sections/Carousel.tsx": $$$1,
-    "./sections/CollectionHighlights.tsx": $$$2,
-    "./sections/Footer.tsx": $$$3,
-    "./sections/Head.tsx": $$$4,
-    "./sections/Header.tsx": $$$5,
-    "./sections/HighlightedFeatures.tsx": $$$6,
+    "./sections/Footer.tsx": $$$2,
+    "./sections/Head.tsx": $$$3,
+    "./sections/Header.tsx": $$$4,
+    "./sections/HighlightedFeatures.tsx": $$$5,
+    "./sections/Highlights.tsx": $$$6,
     "./sections/MarcasConvidadas.tsx": $$$7,
     "./sections/Newsletter.tsx": $$$8,
     "./sections/ProductDetails.tsx": $$$9,
@@ -219,56 +219,6 @@ const manifest: DecoManifest = {
           },
         },
         "required": [],
-      },
-      "outputSchema": null,
-    },
-    "./sections/CollectionHighlights.tsx": {
-      "inputSchema": {
-        "title": " Collection Highlights",
-        "type": "object",
-        "properties": {
-          "highlights": {
-            "type": "array",
-            "items": {
-              "title": "Highlight",
-              "type": "object",
-              "properties": {
-                "src": {
-                  "format": "image-uri",
-                  "type": "string",
-                  "title": "Src",
-                },
-                "alt": {
-                  "type": "string",
-                  "title": "Alt",
-                },
-                "href": {
-                  "type": "string",
-                  "title": "Href",
-                },
-                "label": {
-                  "type": "string",
-                  "title": "Label",
-                },
-              },
-              "required": [
-                "src",
-                "alt",
-                "href",
-                "label",
-              ],
-            },
-            "title": "Highlights",
-          },
-          "title": {
-            "type": "string",
-            "title": "Title",
-          },
-        },
-        "required": [
-          "highlights",
-          "title",
-        ],
       },
       "outputSchema": null,
     },
@@ -481,6 +431,55 @@ const manifest: DecoManifest = {
         },
         "required": [
           "features",
+        ],
+      },
+      "outputSchema": null,
+    },
+    "./sections/Highlights.tsx": {
+      "inputSchema": {
+        "title": " Highlights",
+        "type": "object",
+        "properties": {
+          "highlights": {
+            "type": "array",
+            "items": {
+              "title": "Highlight",
+              "type": "object",
+              "properties": {
+                "src": {
+                  "format": "image-uri",
+                  "type": "string",
+                  "title": "Src",
+                },
+                "alt": {
+                  "type": "string",
+                  "title": "Alt",
+                },
+                "href": {
+                  "type": "string",
+                  "title": "Href",
+                },
+                "label": {
+                  "type": "string",
+                  "title": "Label",
+                },
+              },
+              "required": [
+                "src",
+                "alt",
+                "href",
+                "label",
+              ],
+            },
+            "title": "Highlights",
+          },
+          "title": {
+            "type": "string",
+            "title": "Title",
+          },
+        },
+        "required": [
+          "title",
         ],
       },
       "outputSchema": null,
